Issue Resolved - clock incrementing by 2 seconds
Issue: Incorrect clock behaviour, the clock should increment by 1 seconds and not 2 seconds

Cause: Both the Analog and Digital Clock are palced in a separate component. Both these components makes a call to incClock() and both adds 1 seconds to the current time, resulting in total of 2 seconds.

Solution: The interval subscription and the call to incClock() moved to app.component.ts file now (centralized) and the time is send out to both the Analog and Digital Clock resulting in centralized call to incClock() method adding only 1 second to the time and sending same value of time to both the components, correcting the clock behaviour.